Fix gcc warnings on Windows
git-svn-id: http://ludobloom.com/svn/StemLibProjects@2222 d291a34c-4c4d-0410-8556-a73de5c6a500

+++ jsonserialization/trunk/source/unittest/JSONDeserializationContextTest.c
@@ -143,7 +143,7 @@ static void testInit() {
 	const void * value; \
 	size_t length; \
 	value = context->readBlob(context, KEY, &length); \
-	TestCase_assert(length == EXPECTED_LENGTH, "Expected " SIZE_T_FORMAT " but got " SIZE_T_FORMAT, EXPECTED_LENGTH, length); \
+	TestCase_assert(length == EXPECTED_LENGTH, "Expected " SIZE_T_FORMAT " but got " SIZE_T_FORMAT, (size_t) EXPECTED_LENGTH, length); \
 	TestCase_assert(!memcmp(value, (EXPECTED_VALUE), EXPECTED_LENGTH), "Expected \"%s\" but got \"%s\"", (char *) (EXPECTED_VALUE), (char *) value); \
 	TestCase_assert(context->status == SERIALIZATION_ERROR_OK, "Got error from operation that should have succeeded: %d", context->status); \
 }
